feat: 0906 LMS Produção 2

+++ node/core/middleware/rate-limit.ts
@@ -1,5 +1,5 @@
-import { type Middleware } from "../router.ts";
-import { RouteError } from "../utils/router-error.ts";
+import { type Middleware } from '../router.ts';
+import { RouteError } from '../utils/route-error.ts';
 
 type Request = {
   hits: number;
@@ -9,22 +9,16 @@ type Request = {
 export const rateLimit = (time: number, max: number): Middleware => {
   const requests = new Map<string, Request>();
 
-  setInterval(
-    () => {
-      const now = Date.now();
-
-      for (const [key, request] of requests) {
-        if (now >= request.reset) {
-          requests.delete(key);
-        }
-      }
-    },
-    30 * 60 * 1000,
-  ).unref();
+  setInterval(() => {
+    const now = Date.now();
+    for (const [key, item] of requests) {
+      if (now >= item.reset) requests.delete(key);
+    }
+  }, 30 * 60 * 1000).unref();
 
   return (req, res) => {
-    const key = req.ip;
     const now = Date.now();
+    const key = req.ip;
     let request = requests.get(key);
 
     if (request === undefined || now >= request.reset) {
@@ -32,22 +26,20 @@ export const rateLimit = (time: number, max: number): Middleware => {
         hits: 0,
         reset: now + time,
       };
-
       requests.set(key, request);
     }
 
     request.hits += 1;
 
-    const rLeft = Math.max(0, max - request.hits);
     const sLeft = Math.ceil((request.reset - now) / 1000);
+    const rLeft = Math.max(0, max - request.hits);
     const sTime = Math.ceil(time / 1000);
-
-    res.setHeader("RateLimit", `"deafult"; r=${rLeft}; t=${sLeft}`);
-    res.setHeader("RateLimit-Policy", `"deafult"; q=${max}; w=${sTime}`);
+    res.setHeader('RateLimit', `"default";r=${rLeft};t=${sLeft}`);
+    res.setHeader('RateLimit-Policy', `"default";q=${max};w=${sTime}`);
 
     if (request.hits > max) {
-      res.setHeader("Retry-After", sLeft.toString());
-      throw new RouteError(429, "rate limit exceeded");
+      res.setHeader('Retry-After', `${sLeft}`);
+      throw new RouteError(429, 'rate-limit');
     }
   };
 };
